+++ NexmoClient/Internal/NexmoCore/Network/NXMNetworkManager.m
@@ -202,6 +202,24 @@
           onError:(NXMErrorCallback _Nullable)onError{
     [self.router getEvents:getEventsRequest onSuccess:onSuccess onError:onError];
 }
+
+- (void)getEventsPageWithRequest:(NXMGetEventsPageRequest *)request
+               eventsPagingProxy:(id<NXMPageProxy>)pagingProxy
+                       onSuccess:(void (^)(NXMEventsPage * _Nullable))onSuccess
+                         onError:(void (^)(NSError * _Nullable))onError {
+    [self.router getEventsPageWithRequest:request
+                     eventsPagingProxy:pagingProxy
+                             onSuccess:onSuccess
+                               onError:onError];
+}
+
+- (void)getEventsPageForURL:(NSURL *)url
+          eventsPagingProxy:(id<NXMPageProxy>)proxy
+                  onSuccess:(void (^)(NXMEventsPage * _Nullable))onSuccess
+                    onError:(void (^)(NSError * _Nullable))onError {
+    [self.router getEventsPageForURL:url eventsPagingProxy:proxy onSuccess:onSuccess onError:onError];
+}
+
 - (void)getConversationDetails:(nonnull NSString*)conversationId
                      onSuccess:(NXMSuccessCallbackWithConversationDetails _Nullable)onSuccess
                        onError:(NXMErrorCallback _Nullable)onError {